Moved more logic out of model [skip ci]
