+++ lib/searchkick.rb
@@ -13,6 +13,7 @@ require "searchkick/multi_search"
 require "searchkick/query"
 require "searchkick/reindex_queue"
 require "searchkick/record_data"
+require "searchkick/record_indexer"
 require "searchkick/results"
 require "searchkick/version"
